Consistent usage of usize for offsets

+++ src/pedersen.rs
@@ -17,7 +17,7 @@ pub const SHIFT_POINT: Affine = Affine::Point {
     )),
 };
 
-pub const N_ELEMENT_BITS: u32 = 251;
+pub const N_ELEMENT_BITS: usize = 251;
 
 pub fn hash(elements: &[U256]) -> U256 {
     let mut result = Jacobian::from(SHIFT_POINT);
@@ -28,7 +28,7 @@ pub fn hash(elements: &[U256]) -> U256 {
         let end = start + (N_ELEMENT_BITS as usize);
         for (j, point) in PEDERSEN_POINTS[start..end].iter().enumerate() {
             // TODO: Convert u32 to usize in u256::bit
-            if element.bit(j as u32) {
+            if element.bit(j) {
                 result += point;
             }
         }
